+++ railties/test/generators/actions_test.rb
@@ -200,6 +200,30 @@ class ActionsTest < Rails::Generators::TestCase
     assert_file 'config/routes.rb', /#{Regexp.escape(route_command)}/
   end
 
+  def test_route_should_add_data_with_an_new_line
+    run_generator
+    action :route, "root 'welcome#index'"
+    route_path = File.expand_path("config/routes.rb", destination_root)
+    content = File.read(route_path)
+
+    # Remove all of the comments and blank lines from the routes file
+    content.gsub!(/^  \#.*\n/, '')
+    content.gsub!(/^\n/, '')
+
+    File.open(route_path, "wb") { |file| file.write(content) }
+    assert_file "config/routes.rb", /\.routes\.draw do\n  root 'welcome#index'\nend\n\z/
+
+    action :route, "resources :product_lines"
+
+    routes = <<-F
+Rails.application.routes.draw do
+  resources :product_lines
+  root 'welcome#index'
+end
+F
+    assert_file "config/routes.rb", routes
+  end
+
   def test_readme
     run_generator
     Rails::Generators::AppGenerator.expects(:source_root).times(2).returns(destination_root)